[#120] refactor: 로직 수정에 따른 Controller, Dto, Mapper 수정

+++ server/src/main/java/com/gigker/server/domain/content/dto/ContentApplyResponseDto.java
@@ -26,9 +26,9 @@ public class ContentApplyResponseDto {
 		private String nickName;
 		private String pictureUrl;
 		private String about;
-		private int likeCount;
-		private int dislikeCount;
-		private int reviewCount;
+		private Long likeCount;
+		private Long dislikeCount;
+		private Long reviewCount;
 		private ContentApply.ApplyStatus applyStatus;
 	}
 }
